requirements.txt
certifi==2019.3.9
chardet==3.0.4
docutils==0.14
idna==2.8
Kivy==2.0.0dev0
Kivy-Garden==0.1.4
Pillow==6.0.0
Pygments==2.4.2
requests==2.22.0
urllib3==1.25.3
